@@ -1,9 +1,13 @@
 apply plugin: 'com.android.application'
-apply plugin: 'android-apt'
+
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.3"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 
     packagingOptions
             {
@@ -19,7 +23,7 @@ android {
     defaultConfig {
         applicationId "com.rajit.android"
         minSdkVersion 17
-        targetSdkVersion 25
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -35,21 +39,21 @@ dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
 
 
-    compile('com.mikepenz:materialdrawer:5.3.4@aar') {
-        transitive = true
-    }
-    compile 'com.android.support:appcompat-v7:25.3.1'
+    implementation "com.mikepenz:materialdrawer:6.1.1"
+    compile 'androidx.appcompat:appcompat:1.1.0-alpha01'
     compile 'com.loopj.android:android-async-http:1.4.9'
-    compile 'de.hdodenhof:circleimageview:2.1.0'
+    implementation 'de.hdodenhof:circleimageview:2.2.0'
 
-    compile 'com.android.support:design:25.3.1'
-    compile 'com.applozic.communication.uiwidget:mobicomkitui:4.53'
-    compile 'com.jakewharton:butterknife:8.2.1'
+    compile 'com.google.android.material:material:1.1.0-alpha03'
+    implementation 'com.applozic.communication.uiwidget:mobicomkitui:5.30'
+    implementation 'com.jakewharton:butterknife:10.0.0'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.0.0'
 
-    compile 'com.github.bumptech.glide:glide:3.7.0'
-    compile 'com.google.code.gson:gson:2.7'
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
-    compile 'com.karumi:dexter:4.1.0'
+    implementation 'com.github.bumptech.glide:glide:4.8.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.8.0'
+    implementation 'com.google.code.gson:gson:2.8.5'
+    compile 'androidx.constraintlayout:constraintlayout:2.0.0-alpha3'
+    compile 'com.karumi:dexter:5.0.0'
     testCompile 'junit:junit:4.12'
-    apt 'com.jakewharton:butterknife-compiler:8.2.1'
+    compile 'com.github.CymChad:BaseRecyclerViewAdapterHelper:2.9.46'
 }
