@@ -4,14 +4,15 @@ import android.app.Activity;
 import android.content.Intent;
 import android.net.Uri;
 import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import com.applozic.mobicomkit.uiwidgets.conversation.ConversationUIService;
 import com.applozic.mobicomkit.uiwidgets.conversation.activity.ConversationActivity;
+import com.bumptech.glide.Glide;
+import com.chad.library.adapter.base.BaseQuickAdapter;
+import com.chad.library.adapter.base.BaseViewHolder;
 import com.karumi.dexter.Dexter;
 import com.karumi.dexter.PermissionToken;
 import com.karumi.dexter.listener.PermissionDeniedResponse;
@@ -21,62 +22,45 @@ import com.karumi.dexter.listener.single.PermissionListener;
 import com.rajit.android.R;
 import com.rajit.android.models.User;
 import com.rajit.android.utils.HelperClass;
-import com.squareup.picasso.Picasso;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import de.hdodenhof.circleimageview.CircleImageView;
 
-public class UserListAdapter extends ArrayAdapter<User> {
+public class UserListAdapterAnother extends BaseQuickAdapter<User, BaseViewHolder> {
     private Activity activity;
-
-    public UserListAdapter(Activity activity, User[] users) {
-        super(activity, 0, users);
+    public UserListAdapterAnother(Activity activity, List<User> users) {
+        super(R.layout.user_row, users);
         this.activity = activity;
-
     }
 
     @Override
-    public View getView(int position, View convertView, ViewGroup parent) {
+    protected void convert(BaseViewHolder holder, User user) {
+        holder.setText(R.id.buddy_name, user.getName());
 
-        final User curr_user = getItem(position);
-        if (convertView == null) {
-            convertView = activity
-                    .getLayoutInflater().inflate(R.layout.user_row, parent, false);
-        }
-        TextView userName = convertView.findViewById(R.id.buddy_name);
-        TextView designation = convertView.findViewById(R.id.txtDesignation);
-        Button callButton = convertView.findViewById(R.id.btnCall);
-        Button msgButton = convertView.findViewById(R.id.btnMessage);
-        CircleImageView buddyImage = convertView.findViewById(R.id.buddy_image);
-        designation.setText(curr_user.getDesignation());
-        if (curr_user.getPhone().equals(HelperClass.getPhone(activity))) {
-            userName.setText(curr_user.getName() + "(Me)");
-        } else {
-            userName.setText(curr_user.getName());
-        }
+        holder.setText(R.id.txtDesignation, user.getDesignation());
 
-        if (curr_user.getImageurl() != null) {
-            if (!curr_user.getImageurl().isEmpty())
-                Picasso.get().load(curr_user.getImageurl()).placeholder(getContext().getResources().getDrawable(R.drawable.no_image)).error(getContext().getResources().getDrawable(R.drawable.no_image)).into(buddyImage);
-
-        }
-
-        msgButton.setOnClickListener(new View.OnClickListener() {
+        Button callButton = holder.getView(R.id.btnCall);
+        callButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                if (curr_user.getPhone().equals(HelperClass.getPhone(activity))) {
-                    Toast.makeText(activity, "You can't call yourself!", Toast.LENGTH_LONG).show();
+                if (user.getPhone().equals(HelperClass.getPhone(activity))) {
+                    Toast.makeText(activity, "You can't message yourself!", Toast.LENGTH_LONG).show();
                     return;
                 }
                 Intent intent = new Intent(activity, ConversationActivity.class);
-                intent.putExtra(ConversationUIService.USER_ID, curr_user.getPhone().trim());
-                intent.putExtra(ConversationUIService.DISPLAY_NAME, curr_user.getName().trim()); //put it for displaying the title.
+                intent.putExtra(ConversationUIService.USER_ID, user.getPhone().trim());
+                intent.putExtra(ConversationUIService.DISPLAY_NAME, user.getName().trim()); //put it for displaying the title.
                 intent.putExtra(ConversationUIService.TAKE_ORDER, true);
                 activity.startActivity(intent);
 
             }
         });
 
-        callButton.setOnClickListener(new View.OnClickListener() {
+        Button msgButton = holder.getView(R.id.btnMessage);
+
+        msgButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
 
@@ -86,12 +70,12 @@ public class UserListAdapter extends ArrayAdapter<User> {
 
                             @Override
                             public void onPermissionGranted(PermissionGrantedResponse response) {
-                                if (curr_user.getPhone().equals(HelperClass.getPhone(activity))) {
+                                if (user.getPhone().equals(HelperClass.getPhone(activity))) {
                                     Toast.makeText(activity, "You can't call yourself!", Toast.LENGTH_LONG).show();
                                     return;
                                 }
                                 Intent callIntent = new Intent(Intent.ACTION_CALL);
-                                callIntent.setData(Uri.parse("tel:" + curr_user.getPhone()));
+                                callIntent.setData(Uri.parse("tel:" + user.getPhone()));
                                 activity.startActivity(callIntent);
                             }
 
@@ -111,7 +95,13 @@ public class UserListAdapter extends ArrayAdapter<User> {
             }
         });
 
-        return convertView;
+        CircleImageView buddyImage = holder.getView(R.id.buddy_image);
+
+        if(user.getImageurl() !=null)
+        {
+            if(!user.getImageurl().isEmpty())
+                Glide.with(activity).load(user.getImageurl()).into(buddyImage);
+        }
 
     }
 }
